pyproject.toml
[tool.poetry]
name = "http-forwarding"
version = "0.1.0"
description = ""
authors = ["myuan <provefars@gmail.com>"]
license = "BSD-3"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.62.0"
pymongo = "^3.11.1"
uvicorn = "^0.12.3"
requests = "^2.25.0"
pyfunctional = "^1.4.2"
Jinja2 = "^2.11.2"
aiofiles = "^0.6.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
